Hospital Code
Suppose there was a hospital. We have a line thats first come first serve but the last patient is about to die while the first has a cold. This allows patients to sign up and check out a priority sorted first come first serve list. InputMismatchExceptions are applied in spots needed